+++ src/app/orders/orders.module.ts
@@ -5,9 +5,10 @@ import { SharedModule } from '../shared/shared.module';
 import { OrderItemComponent, OrderListComponent } from './components';
 import { OrdersRoutingModule } from './orders-routing.module';
 import { MakeOrderComponent } from './components/make-order/make-order.component';
+import { OrderFormComponent } from './components/order-form/order-form.component';
 
 @NgModule({
-  declarations: [OrderListComponent, OrderItemComponent, MakeOrderComponent],
+  declarations: [OrderListComponent, OrderItemComponent, MakeOrderComponent, OrderFormComponent],
   imports: [
     SharedModule,
     OrdersRoutingModule